fix: Resolve circular import between rag and tools
Refactors the codebase to eliminate a circular import error that occurred between `src/rag.py` and `src/tools.py`.

The `get_combined_docs` function and its related caching logic have been moved from `src/rag.py` to the more appropriate `src/utils.py`. The `update_docs` logic was also moved and refactored to be stateless.

Both `src/rag.py` and `src/tools.py` have been updated to import this shared functionality from `src/utils.py`, breaking the circular dependency.
